+++ src/data/usecases/add-survey/db-add-survey-protocols.ts
@@ -1,2 +1,2 @@
 export * from '../../../domain/usecases/add-survey'
-export * from '../survey/add-survey-repository'
+export * from '../../protocols/db/survey/add-survey-repository'
