+++ src/infra/db/mongodb/survey/survey-mongo-repository.ts
@@ -1,4 +1,4 @@
-import { AddSurveyRepository } from '../../../../data/usecases/survey/add-survey-repository'
+import { AddSurveyRepository } from '../../../../data/protocols/db/survey/add-survey-repository'
 import { AddSurveyModel } from '../../../../domain/usecases/add-survey'
 import { MongoHelper } from '../helpers/mongo-helper'
